working on norm
