+++ srcs/exec/unset_exec.c
@@ -6,7 +6,7 @@
 /*   By: rgeral <rgeral@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/24 13:53:01 by rgeral            #+#    #+#             */
-/*   Updated: 2022/09/24 14:35:55 by rgeral           ###   ########.fr       */
+/*   Updated: 2022/09/30 21:48:00 by rgeral           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	unset_set_value(t_argmode *args, t_args *d, char   *arg)
 {
-	int len;
-	int i;
-	int j;
+	int		len;
+	int		i;
+	int		j;
 	char	**env_copy;
 
 	i = 0;
@@ -28,9 +28,9 @@ int	unset_set_value(t_argmode *args, t_args *d, char   *arg)
 	while (d->env[j])
 	{
 		if (ft_strncmp(d->env[i], d->needle, ft_strlen(d->needle)) == 0)
-		{
 			j++;
-		}
+		else if (ft_strncmp(d->env[i], arg, ft_strlen(d->env[i])) == 0)
+			j++;
 		env_copy[i] = d->env[j];
 		i++;
 		j++;
@@ -52,10 +52,12 @@ int check_if_set(t_argmode *args, t_args *d, char   *arg)
 	i = 0;
 	d->needle = ft_calloc(ft_strlen(arg), sizeof(char));
 	d->needle = ft_strjoin(arg, "="); 
-	//printf("%s\n", d->needle);
+	printf("%s\n", arg);
 	while (d->env[i])
 	{
 		if (ft_strncmp(d->env[i], d->needle, ft_strlen(d->needle)) == 0)
+			unset_set_value(args, d, arg);
+		else if (ft_strncmp(d->env[i], arg, ft_strlen(arg)) == 0)
 			unset_set_value(args, d, arg);
 		i++;
 	}
@@ -75,10 +77,10 @@ int unset_hub(t_argmode *args, t_args *d)
         i++;
     }
 	i = 0;
-	while (d->env[i])
+/*	while (d->env[i])
 	{
 		printf("%s\n", d->env[i]);
 		i++;	
-	}
+	}*/
     return(0);
 }
